Bug 2020293: Fix stuck subPath unmounts
